Disable information leakage in Tomcat error pages
Configure the ErrorReportValve that’s part of the StandardHost to stop
showing Tomcat server info properties.

This is a security hardening measure that’s in a separate commit to
allow easier cherry-picking the future when we upgrade Tomcat.

See https://tomcat.apache.org/tomcat-8.5-doc/config/valve.html#Error_Report_Valve

The ErrorReportValve is a default valve attached to all StandardHosts
in Tomcat.

Configuring `showReport` and `showServerInfo` means that we don't output
the stack trace or version of the Tomcat server as part of any default
error page rendering.