Enable compression feature of Tomcat
We want the origin to be responsible for compressing responses for
clients that support this feature.

See https://tomcat.apache.org/tomcat-8.5-doc/config/http.html#Standard_Implementation

We are expecting this Docker image to be used behind a reverse-proxy of
some sort. The one we're using immediately does not support compressing
responses. So we're making that the responsibility of Tomcat.

If we (or other consumers of this Docker image) need to change that
decision in the future, we can make it configurable.

But for now, we can hardcode it to be on.