Simplify consumption of this image
Reuse the CMD definition in the parent image. We do not think that
any consumers of this image should need to override it. The purpose of
this image is to run TomEE. TomEE is configured using the existing
environment variables defined in the standard Tomcat launcher script
`catalina.sh`. Configuring the dockerised version via environment
variables should also be the pattern.

Defines CATALINA_OPTS to configure a secure-by-default configuration,
listing assumptions.

Provides guidance on how consumers should customise it.