Paramaterise `secure` property of the HTTP Connector
We need to allow the `secure` property of the HTTP connector to be
overridden to false when running Tomcat on a localhost environment and
it is not behind a reverse-proxy that handles TLS termination.

We are adding a new environment variable to the Dockerfile:

* TOMCAT_CONNECTOR_SECURE = true

This is a secure profile by default, suitable for running in production
behind a proxy that terminates TLS.

We are including this in the `CATALINA_OPTS` environment variable that
the image defines.

```sh
ENV CATALINA_OPTS="$CATALINA_OPTS \
-Dconnector.scheme=$TOMCAT_CONNECTOR_SECURE"
```

The property `connector.secure` is referenced in the server.xml file:

```xml
<Connector
port="8080"
protocol="HTTP/1.1"
secure=${connector.secure}/>
```